Treat GithubTeam as a string value
Restricting choices to only whichever teams are
already in use in the environments repository will
fail any time a new team tries to create a new
namespace.

It is likely that any new team will create a new 
namespace as soon as the team is formed, so we can 
expect some of this kind of failure.

So it doesn't seem worth all the extra code 
required to make this a "choose from this list" 
option.