Add "push to registry" step to github action
The release action creates a release from a tag, but that doesn't
trigger the docker image publishing step, which waits for a release to
be created (via the github UI?).

This commit attempts to combine the two actions so that we get a new
docker image at the same time as the new release of the binary.