Store lower-case team name, when creating NS
This change causes the team name to be lower-cased
before we write the rbac.yaml and variables.tf
files.

This fixes a problem later on, when the user tries
to create an ECR - if they have upper-case letters
in their team name, the ECR name generated by our
terraform module will be rejected.

NB: This change is not tested by the unit tests, 
since those don't cover the relevant part of the
code.