Enable `--skip-version-check` for pipeline use
The automatic update feature of the CLI tool means
that all pipelines which use the tool will break
whenever a new version is released. 

This change allows us to add 
`--skip-version-check` to all pipeline commands, 
so that the pipeline runs correctly regardless of
any new releases of the CLI tool.

NB: Merging this change will break our pipelines
again, but hopefully for the last time.