Remove the `name` flag for serviceaccount
We're going to change the way serviceaccounts are
created, to use the new terraform module instead
of the YAML template.

The terraform module allows the name to be set via
a terraform variable, which will usually be set
outside of the `serviceaccount.tf` file, so trying
to set the name via the CLI would be quite 
complicated. So, I'm planning to remove that
option completely. If people want to change the
name of their serviceaccount, they can easily do
that by setting the terraform variable.

This change removes the name option. A later 
change will switch to creating the serviceaccount
as terraform.