Remove the self updating functionality
Now we have a brew bottle for installing the CLI for macs (and plan to
do the same for snaps and linux), we no longer require the often faulty
self-updating mechanism to keep users up to date. This commit changes
the http get request to a message printed to the users terminal as such:

```bash
➜ ./cloud-platform version
Update required. Current version: 1.12.1, Latest version: 1.12.2

To upgrade the cloud platform cli, run brew update/upgrade or grab the
latest version from
https://api.github.com/repos/ministryofjustice/cloud-platform-cli/releases/latest
```