Create good and bad paths for Bump test
The test runs through a common scenario, mainly testing the Terraform source change.

First we create a test file with an array of bytes. We then pass the new version and the module name (in this case it's "test"). The function should change the "test" version to the string passed. If it doesn't, the `checkModuleChange` will return a false bool and the test will fail.