feat(cluster): :sparkles: Add cluster and snapshot struct and constructors
This allows users to pass a client-go clientset to create a cluster object. This is handy when you want a snapshot of data in your cluster. The snapshot struct will be used to compare cluster states - you could argue that the cluster object can be stored and compared in a different variable/block of memory, but I thought it would be useful to mock and test a snapshot explicitly.