fix(recycle): :bug: Add new var for node being terminated
This bug was found in review. When selecting the oldest node, it uses the same node name up until the cluster is validated, in which a new "oldestNode" is selected. Adding this new var name allows me to use it in the final stdout message about finishing node recycling. Purely aesthetic, but important.