build: :recycle: Use debian bullseye image in place of alpine
To make this dockerfile easier to read and maintain it has been split into three stages. Stage one builds the cli. Two builds kubectl and terraform and stage three pulls them altogether and installs the aws cli. The compromise is that the image is now twice as big (around 700MB)