build(awsprofile): add blank profile file in user space
This is to satisfy the terraform providers requirement to use an aws profile after specifying it from a terraform block. As we do this (from the infrastructure repository), we now need to provide this file or it will fail to auth. Please read the following documentation regarding the requirement: https://registry.terraform.io/providers/hashicorp/aws/latest/docs/guides/version-4-upgrade#changes-to-authentication