fix(decode-secret): add "raw" flag to allow user to print unicode
This commit connects to https://github.com/ministryofjustice/cloud-platform-cli/issues/103 and offers mitigation (not a fix) to a problem where a secret contains a unicode char. We use the stdlib json package to handle a "prettyprint" back to the end user. When doing so, we marshal a raw string into json, which handles UTF-8 only (please see here: https://pkg.go.dev/encoding/json#Marshal). This commit gives the option to specify a `--raw` or `-r` flag to the subcommand and print before marshaling. This loops over all secret values, decodes the base64 and then prints them raw with no formatting applied.