refactor(generate-docs): change global flag into a sub-command
The flag package was interfering with Cobra when trying to specify a global flag first e.g. `cloud-platform --skip-version-check version` would return an error about the --skip-version-check not being recognised. We want global flags to be useful, so this commit changes the generate docs flag to a subcommand of the `cloud-platform` root.