refactor(client): rename struct and add awscredentials
it became clear that we need to have an options struct, containing the region, session and profile aws variables. This will allow us to pass the options onto functions outside the package. With this change requires the need to make the functions and structs in this package descriptive. When renaming the client struct to kube client, I also had to change a number of different function calls.